[Refactor] #130 - PolicyPage를 shared/ui 배럴에 export하고 임포트 경로 정리
