+++ src/shared/ui/policy-page/index.ts
@@ -1,0 +1,2 @@
+export { PolicyPage } from "./policy-page";
+export type { PolicyPageProps } from "./policy-page";
